Went back to XRGB8888

+++ src/ww_screen.c
@@ -3,7 +3,7 @@
 
 #include <string.h>
 
-int ww_screen_init(ww_screen_t* const screen, uint16_t* const pixels, size_t const pitch) {
+int ww_screen_init(ww_screen_t* const screen, uint32_t* const pixels, size_t const pitch) {
     memset(screen, 0, sizeof(*screen));
     screen->canvas.pixels = pixels;
     screen->canvas.pitch = pitch;
